yyx
Requirements
As a User, I want to login.
As a User, I want to add/edit/create employees.
As a User, I want to add/edit/create addresses.
As a User, I want to assign address to employees.
Bonus Tasks
As a User, I want to filter employees based on their addresses.
As a User, I want to view historical deleted employees.
As a User, I want to assign roles to employees.